Revert “Update ScanningViewController.isCameraPaused to be a derived value”
Turns out the management of this state is a bit more fragile than I originally thought. The underlying session is modified asynchronously meaning in some cases the _reading_ of the top-level `paused` value could be incorrect.

We’ll need to take a deeper more thoughtful dive into this in the future if we want to keep the value derived.

Also with this change we rever the `isFocusLocked` since the both changes were necessary to support my new implementation (the one we’re reverting)