Expose AVCaptureDevice authorizationStatus information
This PR partially exposes AVCaptureDevice’s authorizationStatus information to allow consumers to react accordingly.

Until now the only way to tell outside of SCUI’s SDK whether or not camera permissions had been granted was to query AVCaptureDevice directly. We expose a getter for this value inside of CameraManager and also add a hook so we may imperatively react to the starting of a camera session failing due to a denial of camera permissions. The latter case is useful if an application wants to update its UI and direct users to Settings for granting permission immediately after it has been denied.